requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
astroid==2.15.1
    # via pylint
attrs==22.2.0
    # via pytest
black==23.3.0
    # via -r requirements.in
build==0.10.0
    # via pip-tools
click==8.1.3
    # via
    #   black
    #   pip-tools
coloredlogs==15.0.1
    # via -r requirements.in
debugpy==1.6.6
    # via -r requirements.in
dictdiffer==0.9.0
    # via -r requirements.in
dill==0.3.6
    # via pylint
dnspython==2.3.0
    # via pymongo
exceptiongroup==1.1.1
    # via pytest
freezegun==1.2.2
    # via -r requirements.in
humanfriendly==10.0
    # via coloredlogs
humanize==4.6.0
    # via -r requirements.in
ijson==3.2.0.post0
    # via -r requirements.in
iniconfig==2.0.0
    # via pytest
isort==5.12.0
    # via pylint
json-merge-patch==0.2
    # via -r requirements.in
lazy-object-proxy==1.9.0
    # via astroid
mccabe==0.7.0
    # via pylint
mock==5.0.1
    # via -r requirements.in
mypy==1.1.1
    # via -r requirements.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
packaging==23.0
    # via
    #   black
    #   build
    #   pytest
pathspec==0.11.1
    # via black
pip-tools==6.12.3
    # via -r requirements.in
platformdirs==3.2.0
    # via
    #   black
    #   pylint
pluggy==1.0.0
    # via pytest
psycopg2-binary==2.9.5
    # via -r requirements.in
pylint==2.17.1
    # via -r requirements.in
pymongo==4.3.3
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
pytest==7.2.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via freezegun
python-dotenv==1.0.0
    # via -r requirements.in
pywatchman==1.4.1
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
smart-open==6.3.0
    # via -r requirements.in
sqlparse==0.4.3
    # via -r requirements.in
tomli==2.0.1
    # via
    #   black
    #   build
    #   mypy
    #   pylint
    #   pytest
tomlkit==0.11.7
    # via pylint
types-mock==5.0.0.6
    # via -r requirements.in
typing-extensions==4.5.0
    # via
    #   astroid
    #   mypy
wheel==0.40.0
    # via
    #   -r requirements.in
    #   pip-tools
wrapt==1.15.0
    # via astroid
xmltodict==0.13.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
